Fixed game_event_model_equip.equipment_id, now working with new equipment_id system
Former-commit-id: bbc6429046c7755d8cc8d526da63ae603fc01b95

+++ src/server/game/GameEventMgr.h
@@ -64,7 +64,7 @@ struct GameEventData
 struct ModelEquip
 {
     uint32 modelid;
-    uint32 equipment_id;
+    int8 equipment_id;
     uint32 modelid_prev;
     uint32 equipement_id_prev;
 };
